requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langgraph
langchain-google-genai